+++ backend/tests/apps/owasp/models/project_health_metrics_test.py
@@ -1,10 +1,7 @@
-from unittest.mock import patch
-
 import pytest
 from django.core.exceptions import ValidationError
 from django.utils import timezone
 
-from apps.owasp.api.internal.nodes.project_health_stats import ProjectHealthStatsNode
 from apps.owasp.models.project import Project
 from apps.owasp.models.project_health_metrics import ProjectHealthMetrics
 
@@ -119,64 +116,3 @@ class TestProjectHealthMetricsModel:
         metrics.pull_request_last_created_at = self.FIXED_DATE
 
         assert getattr(metrics, field_name) == expected_days
-
-    @patch("apps.owasp.models.project_health_metrics.ProjectHealthMetrics.get_stats")
-    @patch("apps.owasp.models.project_health_metrics.Canvas")
-    @patch("apps.owasp.models.project_health_metrics.Table")
-    @patch("apps.owasp.models.project_health_metrics.TableStyle")
-    @patch("apps.owasp.models.project_health_metrics.BytesIO")
-    def test_generate_overview_pdf(
-        self,
-        mock_bytes_io,
-        mock_table_style,
-        mock_table,
-        mock_canvas,
-        mock_get_stats,
-    ):
-        """Test that the command executes without errors."""
-        metrics_stats = ProjectHealthStatsNode(
-            projects_count_healthy=10,
-            projects_count_unhealthy=5,
-            projects_count_need_attention=3,
-            average_score=75.0,
-            total_contributors=150,
-            total_forks=200,
-            total_stars=300,
-            projects_percentage_healthy=66.67,
-            projects_percentage_need_attention=20.00,
-            projects_percentage_unhealthy=13.33,
-            monthly_overall_scores=[],
-            monthly_overall_scores_months=[],
-        )
-        table_data = [
-            ["Metric", "Value"],
-            ["Healthy Projects", f"{metrics_stats.projects_count_healthy}"],
-            ["Unhealthy Projects", f"{metrics_stats.projects_count_unhealthy}"],
-            ["Need Attention Projects", f"{metrics_stats.projects_count_need_attention}"],
-            ["Average Score", f"{metrics_stats.average_score:.2f}"],
-            ["Total Contributors", f"{metrics_stats.total_contributors:,}"],
-            ["Total Forks", f"{metrics_stats.total_forks:,}"],
-            ["Total Stars", f"{metrics_stats.total_stars:,}"],
-            ["Healthy Projects Percentage", f"%{metrics_stats.projects_percentage_healthy:.2f}"],
-            [
-                "Need Attention Projects Percentage",
-                f"%{metrics_stats.projects_percentage_need_attention:.2f}",
-            ],
-            [
-                "Unhealthy Projects Percentage",
-                f"%{metrics_stats.projects_percentage_unhealthy:.2f}",
-            ],
-        ]
-        mock_get_stats.return_value = metrics_stats
-        ProjectHealthMetrics.generate_overview_pdf()
-        mock_bytes_io.assert_called_once()
-        mock_canvas.assert_called_once_with(mock_bytes_io.return_value)
-        canvas = mock_canvas.return_value
-        mock_table.assert_called_once_with(
-            table_data, colWidths="*", style=mock_table_style.return_value
-        )
-        mock_table_style.assert_called_once()
-        mock_table.return_value.wrapOn.assert_called_once_with(canvas, 400, 600)
-        mock_table.return_value.drawOn.assert_called_once_with(canvas, 100, 570)
-        canvas.showPage.assert_called_once()
-        canvas.save.assert_called_once()
