+++ frontend/.env.example
@@ -7,7 +7,6 @@ NEXT_PUBLIC_IDX_URL=http://localhost:8000/idx/
 NEXT_PUBLIC_IS_PROJECT_HEALTH_ENABLED=true
 NEXT_PUBLIC_RELEASE_VERSION=
 NEXT_PUBLIC_SENTRY_DSN=
-NEXT_PUBLIC_VIEWS_URL=http://localhost:8000/
 NEXT_SENTRY_AUTH_TOKEN=
 NEXT_SERVER_CSRF_URL=http://backend:8000/csrf/
 NEXT_SERVER_DISABLE_SSR=false
