+++ frontend/src/server/fetchMetricsOverviewPDF.ts
@@ -1,9 +1,10 @@
-import { AppError } from 'app/global-error'
-import { VIEWS_URL } from 'utils/credentials'
+import { handleAppError } from 'app/global-error'
+import { API_URL } from 'utils/credentials'
 
 export const fetchMetricsOverviewPDF = async (): Promise<void> => {
   try {
-    const response = await fetch(`${VIEWS_URL}owasp/project-health-metrics/overview/pdf`, {
+    const baseUrl = API_URL.split('/api')[0]
+    const response = await fetch(`${baseUrl}/owasp/project-health-metrics/overview/pdf`, {
       method: 'GET',
       headers: {
         accept: 'application/pdf',
@@ -12,26 +13,27 @@ export const fetchMetricsOverviewPDF = async (): Promise<void> => {
 
     if (!response.ok) {
       const message = `Failed to fetch metrics overview PDF: ${response.status} ${response.statusText}`
-      throw new AppError(response.status, message)
+      handleAppError(new Error(message))
     }
 
     const pdfBlob = await response.blob()
-    if (!pdfBlob) {
-      throw new AppError(500, 'PDF blob is empty or undefined')
+    if (pdfBlob.size === 0) {
+      handleAppError(new Error('PDF blob is empty or undefined'))
     }
     const pdfUrl = window.URL.createObjectURL(pdfBlob)
     const link = document.createElement('a')
     link.href = pdfUrl
-    link.setAttribute('download', 'owasp-project-health-metrics-overview.pdf')
+    const fileName = `owasp_metrics_overview_${new Date().toISOString().split('T')[0]}.pdf`
+    link.setAttribute('download', fileName)
     document.body.appendChild(link)
     link.click()
     document.body.removeChild(link)
+    window.URL.revokeObjectURL(pdfUrl)
   } catch (error) {
-    if (error instanceof AppError) {
-      throw error
-    }
-
-    const message = error?.message || 'Unexpected error while fetching metrics overview PDF'
-    throw new AppError(500, message)
+    handleAppError(
+      error instanceof Error
+        ? error
+        : new Error('An unexpected error occurred while fetching the PDF')
+    )
   }
 }
