+++ frontend/src/utils/credentials.ts
@@ -16,4 +16,3 @@ export const NEXTAUTH_URL = process.env.NEXTAUTH_URL
 export const RELEASE_VERSION = process.env.NEXT_PUBLIC_RELEASE_VERSION
 export const SENTRY_AUTH_TOKEN = process.env.NEXT_SENTRY_AUTH_TOKEN
 export const SENTRY_DSN = process.env.NEXT_PUBLIC_SENTRY_DSN
-export const VIEWS_URL = process.env.NEXT_PUBLIC_VIEWS_URL
